+++ backend/src/main/java/pl/edu/pw/ee/backend/entities/user/courier/CourierRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CourierRepository extends JpaRepository<Courier, Long> {
+public interface CourierRepository extends JpaRepository<Courier, Integer> {
 }
